refactor: move approver-nav guard from Navigator into ActivityReportNavigator preflight
Navigator is a generic shared component (used by ActivityReports,
CommunicationLogs, etc.) and must not contain domain-specific field names.

Move the zero-approver guard out of Navigator.onUpdatePage and into an
AR-specific preFlight() function defined in ActivityReportNavigator, passed
via the existing preFlightForNavigation prop — matching the same pattern
LogFormNavigator already uses.

Changes:
- Navigator/index.js: remove REPORT_STATUSES import and AR-specific
  approvers/calculatedStatus check from onUpdatePage; add comment
  documenting the preFlightForNavigation contract for future contributors
- ActivityReportNavigator.js: add preFlight() that blocks navigation on
  submitted/needs-action reports with zero valid approvers; pass it as
  preFlightForNavigation prop to Navigator
- Navigator test: replace AR-specific test with generic preflight-blocks-nav test
- ActivityReportNavigator test: add the AR-specific zero-approver case
